+++ src/methods.ts
@@ -1,4 +1,0 @@
-
-const Methods = [ "GET", "PATCH", "PUT", "POST", "DELETE" ] as const;
-
-export { Methods };