+++ src/operations.ts
@@ -1,0 +1,4 @@
+
+const Operations = [ "select", "insert", "update", "merge", "delete" ] as const;
+
+export { Operations };